requirements.txt
twilio==5.4.0
mock>=1.0
djangorestframework>=3.1.3
djangorestframework-simplejwt[crypto]
coverage
